+++ Changelog.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Apply session.destroy function into deskelectronjs engine. Completed on 2025-08-07
+
 ### Changed
 
 - Redesign data passing from frontend request to backend and date return at deskelectronjs engine. Completed on 2025-07-25